requirements.txt
opencv-python==4.5.5.64
numpy==1.22.3
matplotlib==3.5.1
PyYAML==6.0
pytest==7.1.1
pandas==1.4.2
Pillow==9.1.1
tensorflow==2.9.1
azureml-core==1.48.0
azureml-dataset-runtime==1.43.0
tqdm==4.64.0
ipykernel==6.15.1
yolov5==6.1.8
mlflow==1.29.0
azureml-mlflow==1.46.0
lxml==4.9.2
html5lib==1.1
beautifulsoup4==4.11.1
ultralytics==8.0.42